deployment/requirements.txt
google-cloud-dataproc==5.9.0
google-cloud-storage==2.14.0
google-cloud-core==2.4.1
google-cloud-pubsub==2.19.1
google-cloud-scheduler==2.13.1